requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
openai==1.51.2
numpy==1.26.4
beautifulsoup4==4.12.3    # для html→текст (если потребуется)

